Movendo a implementação do AspnetUser para o shared kernel para o reaproveitamento
