--- src/web/NSE.WebApp.MVC/Extensions/IUser.cs
@@ -1,69 +1,0 @@
-﻿using System.Security.Claims;
-
-namespace NSE.WebApp.MVC.Extensions
-{
-    public interface IUser
-    {
-        string Name { get; }
-        Guid GetUserId();
-        string GetUserEmail();
-        string GetUserToken();
-        bool IsAuthenticated();
-        bool IsInRole(string role);
-        IEnumerable<Claim> GetClaims();
-        HttpContext GetHttpContext();
-    }
-
-    public class AspNetUser : IUser
-    {
-        private readonly IHttpContextAccessor _accessor;
-
-        public AspNetUser(IHttpContextAccessor accessor) => _accessor = accessor;
-
-        public string Name => _accessor.HttpContext.User.Identity.Name;
-
-        public Guid GetUserId() => IsAuthenticated() ? Guid.Parse(_accessor.HttpContext.User.GetUserId()) : Guid.Empty;
-
-        public string GetUserEmail() => IsAuthenticated() ? _accessor.HttpContext.User.GetUserEmail() : string.Empty;
-
-        public string GetUserToken() => IsAuthenticated() ? _accessor.HttpContext.User.GetUserToken() : string.Empty;
-
-        public bool IsAuthenticated() => _accessor.HttpContext.User.Identity.IsAuthenticated;
-
-        public bool IsInRole(string role) => _accessor.HttpContext.User.IsInRole(role);
-
-        public IEnumerable<Claim> GetClaims() => _accessor.HttpContext.User.Claims;
-
-        public HttpContext GetHttpContext() => _accessor.HttpContext;
-    }
-
-    public static class ClaimsPrincipalExtensions
-    {
-        public static string GetUserId(this ClaimsPrincipal principal)
-        {
-            if (principal == null)
-                throw new ArgumentNullException(nameof(principal));
-
-            var claim = principal.FindFirst("sub");
-            return claim?.Value;
-        }
-
-        public static string GetUserEmail(this ClaimsPrincipal principal)
-        {
-            if (principal == null)
-                throw new ArgumentNullException(nameof(principal));
-
-            var claim = principal.FindFirst("email");
-            return claim?.Value;
-        }
-
-        public static string GetUserToken(this ClaimsPrincipal principal)
-        {
-            if (principal == null)
-                throw new ArgumentNullException(nameof(principal));
-
-            var claim = principal.FindFirst("JWT");
-            return claim?.Value;
-        }
-    }
-}
