+++ src/web/NSE.WebApp.MVC/Views/Shared/_LoginPartial.cshtml
@@ -1,6 +1,7 @@
 ﻿@using Microsoft.AspNetCore.Identity
+@using NSE.WebApi.Core.User.Interfaces;
 @using NSE.WebApp.MVC.Extensions
-@inject IUser AppUser
+@inject IAspNetUser AppUser
 
 @if (AppUser.IsAuthenticated())
 {
